Harmonized with wisski_view(); small layout change
When used with the wisski module versions fbb0f44f66 and newer, the theme
would occassionally duplicate wisski content. This fix should work with older
versions, the former ones and newest version of wisski_view().

Made small changes to the layout of non-individual nodes so that display will
not produce white areas between title and text.